requirements.txt
sounddevice==0.5.2
websockets==12.0
twilio==8.10.0
python-dotenv==1.0.0
aiohttp==3.9.1
numpy==1.26.0
